requirements.txt
aiosqlite==0.22.1
anitya_schema==2.3.0
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
Automat==25.4.16
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
constantly==23.10.4
crochet==2.1.1
cryptography==46.0.3
decorator==5.2.1
defusedxml==0.7.1
distro==1.9.0
Django==6.0
djangorestframework==3.16.1
docstring_parser==0.17.0
dotenv==0.9.9
fedora_messaging==3.9.0
fedora_messaging_the_new_hotness_schema==1.3.0
future==1.0.0
gssapi==1.10.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
koji==1.35.3
langchain-core==1.2.20
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
load==1.0.14
load-dotenv==0.1.0
mysqlclient==2.2.8
orjson==3.11.7
ormsgpack==1.12.2
packaging==25.0
pika==1.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-bugzilla==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
redis==5.3.1
referencing==0.37.0
requests==2.32.5
requests-gssapi==1.4.0
requests-toolbelt==1.0.0
rpds-py==0.30.0
service-identity==24.2.0
six==1.17.0
sniffio==1.3.1
sqlite-vec==0.1.7
sqlparse==0.5.5
tenacity==9.1.4
the-new-hotness==1.3.0
tomli==2.3.0
Twisted==25.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.14.1
wrapt==2.0.1
xxhash==3.6.0
zope.interface==8.1.1
zstandard==0.25.0
